requirements.txt
annotated-types==0.7.0
anyio==4.11.0
appdirs==1.4.4
attrs==25.3.0
beautifulsoup4==4.13.4
browser-cookie3==0.20.1
bs4==0.0.2
certifi==2025.8.3
charset-normalizer==3.4.2
contourpy==1.3.3
cssselect==1.3.0
cycler==0.12.1
dateparser==1.2.2
demjson3==3.0.6
distro==1.9.0
et_xmlfile==2.0.0
facebook-scraper==0.2.59
fake-useragent==2.2.0
fonttools==4.60.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
jiter==0.11.1
joblib==1.5.2
kiwisolver==1.4.9
lxml==6.0.0
lxml_html_clean==0.4.2
lz4==4.4.4
matplotlib==3.10.7
numpy==2.3.4
openai==2.6.1
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.3
parse==1.20.2
pillow==12.0.0
pyaes==1.6.1
pyasn1==0.6.1
pycryptodomex==3.23.0
pydantic==2.12.3
pydantic_core==2.41.4
pyee==11.1.1
pyparsing==3.2.5
pyppeteer==2.0.0
pyquery==2.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2025.2
regex==2025.7.34
requests==2.32.4
requests-html==0.10.0
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.16.3
seaborn==0.13.2
selenium==4.34.2
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
tabulate==0.9.0
Telethon==1.36.0
threadpoolctl==3.6.0
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.2
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
w3lib==2.3.1
websocket-client==1.8.0
websockets==10.4
wsproto==1.2.0
zipp==3.23.0
